ADD: Conda environment config for running the supplementary local scripts
The notebook for generating the dataset train-test split and the setup and initialization of the FAISS index of the GitHub issue descriptions are performed locally, outside the Docker container. To facilitate reproducibility we provide an anaconda environment. We need an anaconda environment, because FAISS only support easy installs from conda.